trade-model/requirements.txt
# Core ML libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
joblib>=1.3.0
python-dotenv>=1.0.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and notebooks
jupyterlab>=4.0.0
ipykernel>=6.25.0

# Optional: for advanced features
ta-lib>=0.4.25  # Technical analysis library
yfinance>=0.2.18  # For backup data source
